Quick fix for commit r538551(DRLVM build fails on Linux): (Applied HARMONY-3636 [DRLVM][jit]Color-graph register allocator enable in server mode)
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@538610 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/codegenerator/ia32/Ia32RegAlloc3.cpp
@@ -332,7 +332,7 @@ private:
 
     char& at (int i, int j) const
     {
-        assert(0 <= i && i < dim && 0 <= j && j < dim && i != j);
+        assert(0 <= i && i < (int)dim && 0 <= j && j < (int)dim && i != j);
 
         return (i < j) ? *(i*dim + j + ptr)
                        : *(j*dim + i + ptr);
